backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
transformers>=4.45.0
tokenizers>=0.22.0,<=0.23.0
huggingface_hub>=0.23.0
accelerate
opencv-python
pillow
numpy
openai
moviepy
timm
packaging